+++ src/main/java/com/project/KiTucXa/Dto/Response/PaymentResponse.java
@@ -3,6 +3,7 @@ package com.project.KiTucXa.Dto.Response;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 @Data
@@ -13,7 +14,7 @@ import java.sql.Timestamp;
 public class PaymentResponse {
     String paymentId;
     String billId;
-    int installmentCount; // số lần đóng (tối đa ba lần)
+    BigDecimal installmentCount; // số lần đóng (tối đa ba lần)
     String note;
     Timestamp createdAt;
     Timestamp updatedAt;
